# pruebas/openapi.yaml
openapi: 3.1.0
info:
  title: Billing Backend System
  version: 0.1.0
paths:
  /auth/register:
    post:
      tags:
      - Autenticación
      summary: Register User
      operationId: register_user_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/app__schemas__auth_schemas__UsuarioCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/login:
    post:
      tags:
      - Autenticación
      summary: Login de usuario
      operationId: login_user_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioLogin'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    UsuarioLogin:
      properties:
        usuario_email:
          type: string
          format: email
          title: Usuario Email
        usuario_password:
          type: string
          title: Usuario Password
      type: object
      required:
      - usuario_email
      - usuario_password
      title: UsuarioLogin
    UsuarioResponse:
      properties:
        usuario_email:
          type: string
          title: Usuario Email
      type: object
      required:
      - usuario_email
      title: UsuarioResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    app__schemas__auth_schemas__UsuarioCreate:
      properties:
        usuario_email:
          type: string
          format: email
          title: Usuario Email
        usuario_password:
          type: string
          title: Usuario Password
        nombre:
          type: string
          title: Nombre
        apellido:
          type: string
          title: Apellido
      type: object
      required:
      - usuario_email
      - usuario_password
      - nombre
      - apellido
      title: UsuarioCreate
...